Refactor firewall class (#1711)
* wip

* separated account interfaces

* refactored account methods

* fix build (#1706)

* fix build

* missing semi colon

* wip

* wip

* fixed unit tests

* refactored firewall and account

* fixed linter issue

* moved signin flag, decoupled passing session logic

* wip

* fixed test

* fixed linter issue

* fix

* fix
