+++ test/accountService.test.ts
@@ -5,17 +5,15 @@
 import * as vscode from 'vscode';
 import * as TypeMoq from 'typemoq';
 import SqlToolsServiceClient from '../src/languageservice/serviceclient';
-import { FirewallService } from '../src/firewall/firewallService';
 import { AccountService } from '../src/azure/accountService';
 import { HandleFirewallRuleRequest, IHandleFirewallRuleResponse,
     CreateFirewallRuleRequest, ICreateFirewallRuleResponse, IHandleFirewallRuleParams } from '../src/models/contracts/firewall/firewallRequest';
 import VscodeWrapper from '../src/controllers/vscodeWrapper';
 import { assert } from 'chai';
-import { IAzureSession, IAzureResourceFilter } from '../src/models/interfaces';
+import { IAzureSession } from '../src/models/interfaces';
 
 
 suite('Firewall Service Tests', () => {
-    let firewallService: FirewallService;
     let accountService: AccountService;
     let client: TypeMoq.IMock<SqlToolsServiceClient>;
     let vscodeWrapper: TypeMoq.IMock<VscodeWrapper>;
@@ -40,10 +38,6 @@ suite('Firewall Service Tests', () => {
             tenantId: 'test',
             credentials: undefined
         };
-        let mockFilter: IAzureResourceFilter = {
-            session: mockSession,
-            subscription: undefined
-        };
         let mockExtension: vscode.Extension<any> = {
             id: '',
             extensionKind: undefined,
@@ -53,35 +47,11 @@ suite('Firewall Service Tests', () => {
             activate: undefined,
             extensionUri: undefined,
             exports: {
-                filter: [mockFilter]
+                sessions: [mockSession]
             }
         };
         vscodeWrapper.setup(v => v.azureAccountExtension).returns(() => mockExtension);
         accountService = new AccountService(client.object, vscodeWrapper.object);
-        firewallService = new FirewallService(accountService);
     });
 
-
-
-    test('Handle Firewall Rule test', async () => {
-        let result = await firewallService.handleFirewallRule(12345, 'firewall error!');
-        assert.isNotNull(result, 'Handle Firewall Rule request is sent successfully');
-    });
-
-    test('Create Firewall Rule Test', async () => {
-        let server = 'test_server';
-        let startIpAddress = '1.2.3.1';
-        let endIpAddress = '1.2.3.255';
-        accountService.initializeSessionAccount();
-        let mockToken = {
-            expiresOn: new Date(),
-            resource: undefined,
-            tokenType: 'test',
-            accessToken: 'test_token'
-        };
-        accountService.token = mockToken;
-        let result = await firewallService.createFirewallRule(server, startIpAddress, endIpAddress);
-        assert.isNotNull(result, 'Create Firewall Rule request is sent successfully');
-    });
 });
-
